Role added, unified index.html

+++ src/main/java/ua/cafe/controllers/DishController.java
@@ -12,6 +12,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import ua.cafe.entities.Dish;
 import ua.cafe.entities.JsonMaker;
+import ua.cafe.entities.Role;
 import ua.cafe.services.CategoriesService;
 import ua.cafe.services.DishService;
 import ua.cafe.services.UserService;
@@ -46,10 +47,9 @@ public class DishController {
 
     @GetMapping("/menuJSON")
     public ResponseEntity<String> getMenuJson() {
-        String json;
         List<Dish> dishes = dishService.getAllDishes();
         for (Dish dish : dishes) {
-            dish.setDetails(null);
+            dish.setDetails(null); //to avoid cyclic referencing in JSON [dish->details->dish...]
         }
         return JsonMaker.getJsonResponse(dishes);
     }
@@ -59,8 +59,11 @@ public class DishController {
         return JsonMaker.getJsonResponse(dishService.getById(id));
     }
 
-    @PostMapping("/dish_update")
-    public ResponseEntity<String> updateDish(@Valid Dish dish, BindingResult result) {
+    @PostMapping("/dish_update_direct")
+    public ResponseEntity<String> updateDish(@Valid Dish dish, BindingResult result, Principal principal) {
+        Role role = new Role(principal);
+        if (!role.isAdmin())
+            return new ResponseEntity<>("You have no permission to change dishes.",HttpStatus.FORBIDDEN);
         if (result.hasErrors()) {
             return ResponseEntity.status(HttpStatus.NOT_ACCEPTABLE).build();
         }
@@ -70,31 +73,41 @@ public class DishController {
     }
 
     @GetMapping("/menu")
-    public String getMenu(Model model, Principal principal ) {
-        try{
+    public String getMenu(Model model, Principal principal) {
+        try {
             model.addAttribute("dishes", dishService.getAllDishes());
             final UserDetails user = userService.loadUserByUsername(principal.getName());
 
-            switch(user.getAuthorities().toString()) {
-                case "[ROLE_WAITER]":  return "Waiter/menu";
-                case "[ROLE_COOK]":  return "Cook/menu";
-                case "[ROLE_ADMIN]":  return "Director/menu";
+            switch (user.getAuthorities().toString()) {
+                case "[ROLE_WAITER]":
+                    return "Waiter/menu";
+                case "[ROLE_COOK]":
+                    return "Cook/menu";
+                case "[ROLE_ADMIN]":
+                    return "Director/menu";
             }
         } catch (NullPointerException e) {
             return "User/menu";
         }
         return "User/menu";
     }
-////////////////////////
+
+    ////////////////////////
     @GetMapping("/add_dish")
-    public String addDish(Model model) {
+    public String addDish(Model model, Principal principal) {
+        Role role = new Role(principal);
+        if (!role.isAdmin())
+            return "/permissionDenied";
         model.addAttribute("dish", new Dish());
         model.addAttribute("categories", categoriesService.getAllCategories());
         return "Director/add_dish";
     }
 
     @GetMapping("/dish_edit")
-    public String editDish(Model model, @RequestParam Long id) {
+    public String editDish(Model model, @RequestParam Long id, Principal principal) {
+        Role role = new Role(principal);
+        if (!role.isAdmin())
+            return "/permissionDenied";
         Dish dish = dishService.getById(id);
         model.addAttribute("dish", dish);
         model.addAttribute("categories", categoriesService.getAllCategories());
@@ -102,8 +115,11 @@ public class DishController {
         return "Director/edit_dish";
     }
 
-    @PostMapping("/dish_update1")
-    public String updateDish1(@Valid Dish dish, BindingResult result, Model model) {
+    @PostMapping("/dish_update")
+    public String updateDish1(@Valid Dish dish, BindingResult result, Model model, Principal principal) {
+        Role role = new Role(principal);
+        if (!role.isAdmin())
+            return "/permissionDenied";
         if (result.hasErrors()) {
             model.addAttribute("categories", categoriesService.getAllCategories());
             return "Director/edit_dish";
@@ -114,7 +130,10 @@ public class DishController {
     }
 
     @PostMapping("/add_dish")
-    public String greetingSubmit(@Valid Dish dish, BindingResult result, Model model) {
+    public String greetingSubmit(@Valid Dish dish, BindingResult result, Model model, Principal principal) {
+        Role role = new Role(principal);
+        if (!role.isAdmin())
+            return "/permissionDenied";
         if (result.hasErrors()) {
             model.addAttribute("categories", categoriesService.getAllCategories());
             return "Director/add_dish";
@@ -124,7 +143,10 @@ public class DishController {
     }
 
     @GetMapping("/dish_remove")
-    public String removeDish(@RequestParam Long id) {
+    public String removeDish(@RequestParam Long id, Principal principal) {
+        Role role = new Role(principal);
+        if (!role.isAdmin())
+            return "/permissionDenied";
         if (!dishService.existsWithId(id))
             throw new NoSuchElementException();
         dishService.removeById(id);
